+++ src/main/java/com/bookspot/batch/library/reader/crawler/NaruDetailParser.java
@@ -1,4 +1,4 @@
-package com.bookspot.batch.library.naru;
+package com.bookspot.batch.library.reader.crawler;
 
 import com.bookspot.batch.global.crawler.common.CrawlingResult;
 import com.bookspot.batch.global.crawler.common.JsoupCrawler;
